requirements.txt
asgiref==3.8.1
boto3==1.35.10
botocore==1.35.10
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
defusedxml==0.8.0rc2
Django==5.0.4
django-allauth==64.2.1
django-cors-headers==4.3.1
django-debug-toolbar==4.3.0
django-environ==0.11.2
django-storages==1.14.4
djangorestframework==3.15.1
environ==1.0
idna==3.7
jmespath==1.0.1
oauthlib==3.2.2
pillow==10.2.0
psycopg2==2.9.9
pycparser==2.22
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python3-openid==3.2.0
razorpay==1.4.2
requests==2.31.0
requests-oauthlib==2.0.0
s3transfer==0.10.2
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.3
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.1
whitenoise==6.7.0